keep old controller make resource controller
