+++ server/src/Controllers/Games.ts
@@ -5,9 +5,9 @@ import {
 } from "../db/schema.js";
 import { Request, Response } from "express";
 import GameModel from "../Models/Game.js";
-import { Controller } from "./Controller.js";
+import { ResourceController } from "./ResourceController.js";
 
-export default class GamesController extends Controller {
+export default class GamesController extends ResourceController {
   constructor() {
     super(GameModel, gamesSelectSchema, gamesInsertSchema, gamesUpdateSchema);
   }
